requirements.txt
typer
textual
httpx
pyyaml
rich
asyncio
python-dotenv
pydantic
numpy
duckduckgo-search
